added cli utilities like logging and click commands to _cli.py and used these options in the MCP CLI
